% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_ReadGroups_with_picard.R
\name{add_ReadGroups_with_picard}
\alias{add_ReadGroups_with_picard}
\title{Add Read Groups to SAM files using Picard tools}
\usage{
add_ReadGroups_with_picard(input_BAM, output_BAM, read_group_lib,
  read_group_platform, read_group_barcode, read_group_sample)
}
\arguments{
\item{input_BAM}{specify the input BAM file}

\item{output_BAM}{specify the output BAM file}

\item{read_group_lib}{set the read group library value with a character}

\item{read_group_platform}{set the read group platform value with a character}

\item{read_group_barcode}{set the read group barcode value with a character}

\item{read_group_sample}{set the read group sample value with a character}
}
\description{
Add Read Groups to SAM files using Picard tools
}

