% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_SNPs_with_GATK.R
\name{call_SNPs_with_GATK}
\alias{call_SNPs_with_GATK}
\title{call SNPs with GATK}
\usage{
call_SNPs_with_GATK(ref_fasta, bam_with_rg, Threads, output_vcf, java_mem)
}
\arguments{
\item{ref_fasta}{reference genome or contigs (in fasta format)}

\item{bam_with_rg}{BAM input file with full headers}

\item{Threads}{number of threads to run (depends on server resources)}

\item{output_vcf}{specify the output file (in VCF format)}

\item{java_mem}{set memory usage for SNP calling using GATK}
}
\description{
call SNPs with GATK
}

