% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_RADseqR_project.R
\name{create_RADseqR_project}
\alias{create_RADseqR_project}
\title{create RADseqR project}
\usage{
create_RADseqR_project(this_directory, project_prefix)
}
\arguments{
\item{this_directory}{full path to the directory for creating a RADseqR project}

\item{project_prefix}{prefix name for this RADseqR project}
}
\value{
directories created for the project
}
\description{
create RADseqR project
}

