% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_sequence_dict_with_picard.R
\name{create_sequence_dict_with_picard}
\alias{create_sequence_dict_with_picard}
\title{Create Fasta Dictionary}
\usage{
create_sequence_dict_with_picard(ref_fasta)
}
\arguments{
\item{ref_fasta}{reference fasta or genome}
}
\value{
"Created Sequence Dictionary using PicardTools!"
}
\description{
Create Fasta Dictionary
}

