% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/execute_sort_read_pairs_from_stacks.R
\name{execute_sort_read_pairs_from_stacks}
\alias{execute_sort_read_pairs_from_stacks}
\title{Execute sort read pairs from stacks}
\usage{
execute_sort_read_pairs_from_stacks(stacks_output_files, sample_files,
  whitelist_file, output_files)
}
\arguments{
\item{stacks_output_files}{output files from stacks (from the results of execute_denovomap.pl)}

\item{sample_files}{sample files directory}

\item{whitelist_file}{which whitelist to use}

\item{where}{to store the output files}
}
\description{
Execute sort read pairs from stacks
}

