% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/execute_velvet_script_from_stacks.R
\name{execute_velvet_script_from_stacks}
\alias{execute_velvet_script_from_stacks}
\title{Execute Velvet from Stacks}
\usage{
execute_velvet_script_from_stacks(sort_read_pairs_output, output_dir,
  min_contig_size)
}
\arguments{
\item{sort_read_pairs_output}{output from sort_read_pairs from stacks}

\item{output_dir}{output directory}

\item{min_contig_size}{miniumum size of contigs. Note: this is sensitive to sample read size}
}
\description{
Execute Velvet from Stacks
}

