% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_fastq_to_fasta.R
\name{fastq_to_fasta}
\alias{fastq_to_fasta}
\title{fastq to fasta}
\usage{
fastq_to_fasta(fastq_infile, fasta_outfile)
}
\arguments{
\item{fastq_infile}{file name of fastq input file as a character vector}

\item{fasta_outfile}{file name of fasta output file as a character vector}
}
\description{
fastq to fasta
}

