% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_artifacts_filter.R
\name{fastx_artifacts_filter}
\alias{fastx_artifacts_filter}
\title{FASTX Artifacts Filter}
\usage{
fastx_artifacts_filter(fastq_infile, fastq_outfile)
}
\arguments{
\item{fastq_infile}{name of fastq file as a character vector}

\item{fastq_outfile}{name of fasta file as a character vector}
}
\description{
FASTX Artifacts Filter
}

