% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_clipper.R
\name{fastx_clipper}
\alias{fastx_clipper}
\title{fastx clipper}
\usage{
fastx_clipper(adapter, fastq_infile, fastq_outfile)
}
\arguments{
\item{adapter}{of sample as a character}

\item{fastq_infile}{name of fastq file as a character vector}

\item{fastq_outfile}{name of fasta file as a character vector}
}
\description{
fastx clipper
}

