% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_quality_stats.R
\name{fastx_quality_stats}
\alias{fastx_quality_stats}
\title{fastx quality report}
\usage{
fastx_quality_stats(fastq_infile, text_outfile)
}
\arguments{
\item{fastq_infile}{name of fastq file as a character vector}

\item{fasta_outfile}{name of fasta file as a character vector}
}
\description{
fastx quality report
}

