% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_renamer.R
\name{fastx_renamer}
\alias{fastx_renamer}
\title{fastx renamer}
\usage{
fastx_renamer(fastq_infile, fastq_outfile, n_type)
}
\arguments{
\item{fastq_infile}{name of fastq file as a character vector}

\item{fastq_outfile}{name of fasta file as a character vector}

\item{n_type}{possible values are SEQ and COUNT}
}
\description{
fastx renamer
}

