% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_reverse_complement.R
\name{fastx_reverse_complement}
\alias{fastx_reverse_complement}
\title{FASTX Reverse Complement}
\usage{
fastx_reverse_complement(fastq_infile, fastq_outfile)
}
\arguments{
\item{fastq_infile}{name of fastq file as a character vector}

\item{fastq_outfile}{name of fasta file as a character vector}
}
\description{
FASTX Reverse Complement
}

