% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastx_trimmer.R
\name{fastx_trimmer}
\alias{fastx_trimmer}
\title{FASTX Trimmer}
\usage{
fastx_trimmer(first_base, last_base, fastq_infile, fastq_outfile)
}
\arguments{
\item{first_base}{which first base to keep}

\item{last_base}{which last base to keep}

\item{fastq_infile}{name of fastq file as a character vector}

\item{fastq_outfile}{name of fasta file as a character vector}
}
\description{
FASTX Trimmer
}

