% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_whitelist.R
\name{get_whitelist_from_haplotype_list}
\alias{get_whitelist_from_haplotype_list}
\title{Get whitelist from haplotype list}
\usage{
get_whitelist_from_haplotype_list(haplotypes_file, whitelist_file)
}
\arguments{
\item{haplotype_file}{filename of your haplotypes file. Usually found in your Stacks folder.}

\item{number_of_Mismatches}{maximum number of mismatches}
}
\description{
Get whitelist from haplotype list
}

