% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_this_application_installed.R
\name{is_the_application_installed}
\alias{is_the_application_installed}
\title{Is this application installed on this system ?}
\usage{
is_the_application_installed(the_command)
}
\arguments{
\item{the_command}{application name}
}
\value{
returns OK or NOT FOUND!
}
\description{
Is this application installed on this system ?
}

