% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_psuedoparent.R
\name{make_psuedoparent}
\alias{make_psuedoparent}
\title{make psuedo-parent files}
\usage{
make_psuedoparent(input_samples, psuedoparent)
}
\arguments{
\item{input_samples}{demultiplexed samples to create psuedoparent (as a character list)}

\item{psuedoparent}{desired name of psuedoparent file}
}
\description{
make psuedo-parent files
}

