% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_with_bowtie2.R
\name{map_with_bowtie2}
\alias{map_with_bowtie2}
\title{Map reads with Bowtie2}
\usage{
map_with_bowtie2(ref_fasta, idx_name, fastq_1, fastq_2, Threads, sam_outfile,
  rg_id, rg_lb, rg_sm, rg_pl)
}
\arguments{
\item{ref_fasta}{reference genome or contigs in fasta format}

\item{idx_name}{index name for alignment}

\item{fastq_1}{left or 1 reads}

\item{fastq_2}{left or 2 reads (for pair-end reads only)}

\item{Threads}{number of concurrent threads to run with (depends on server resources)}

\item{sam_outfile}{resulting filename (in SAM format)}

\item{rg_id}{sample id name}

\item{rg_lb}{sample library name}

\item{rg_sm}{sample pool name}

\item{rg_pl}{Platform/technology used to produce reads. Valid values: "CAPILLARY", "LS454", "ILLUMINA", "SOILD","HELICOS","IONTORRENT","ONT","PACBIO"}
}
\description{
Map reads with Bowtie2
}

