% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_with_bwa_and_remove_PCR_duplicates.R
\name{map_with_bwa_and_remove_PCR_duplicates}
\alias{map_with_bwa_and_remove_PCR_duplicates}
\title{Remove PCR duplicates using bwa and Picard}
\usage{
map_with_bwa_and_remove_PCR_duplicates(idxbase, in1_fq, in2_fq, threads,
  output_filename)
}
\arguments{
\item{idxbase}{name of bwa index}

\item{in1_fq}{left reads or 1 reads (of pair-end reads) fastq file}

\item{in2_fq}{right or 2 reads (of pair-end reads) fastq file}

\item{threads}{number of threads to run with (depends on server resources)}

\item{output_filename}{output filename (after removing PCR duplicates)}
}
\description{
Remove PCR duplicates using bwa and Picard
}

