% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_with_BWA_mem.R
\name{map_with_bwa_mem}
\alias{map_with_bwa_mem}
\title{Map reads with BWA}
\usage{
map_with_bwa_mem(ref_fasta, fastq_1, fastq_2)
}
\arguments{
\item{ref_fasta}{reference genome or contigs}

\item{fastq_1}{left reads or 1 reads of sample (pair-end reads) to map}

\item{fastq_2}{right reads or 2 reads of sample (pair-end reads) to map}
}
\description{
Map reads with BWA
}

