% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_ddradtags_from_stacks.R
\name{process_ddradtags_from_stacks}
\alias{process_ddradtags_from_stacks}
\title{Process radtags with Stacks}
\usage{
process_ddradtags_from_stacks(path_to_input_files, barcode_file, output_dir,
  enzyme1, enzyme2, phred_value)
}
\arguments{
\item{path_to_input_files}{Absolute path to input files (as a character)}

\item{barcode_file}{File name of a plain text file with all barcodes for samples (as a character)}

\item{output_dir}{Absolute path for output files (as a character)}

\item{enzyme1}{1st Restriction enzyme used for the assay (as a character)}

\item{enzyme2}{2nd Restriction enzyme used for the assay (as a character)}
}
\description{
Process radtags with Stacks
}

