% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/execute_process_radtags.R
\name{process_radtags_from_stacks}
\alias{process_radtags_from_stacks}
\title{Process radtags with Stacks}
\usage{
process_radtags_from_stacks(path_to_input_files, barcode_file, output_dir,
  enzyme, phred_value)
}
\arguments{
\item{path_to_input_files}{Absolute path to input files (as a character class)}

\item{barcode_file}{File name of a plain text file with all barcodes for samples (as a character class)}

\item{output_dir}{Absolute path for output files (as a character class)}

\item{enzyme}{Restriction enzyme used for the assay (as a character class)}
}
\value{
Process RADseq tags using the process_radtags pipeline from Stacks
}
\description{
Process radtags with Stacks
}

